macOS_Mojave_FullUpdate
This script is designed to perform a full upgrade from a previous version of macOS to Mojave. In this case 10.14.5.
It is deigned to be run once a day prompting a logged in user to install the upgrade. On the 4th attempt a client is forcibly upgraded.
This script is depending on a cached Mojave pkg package.